Functions for PTS are separated in new file
