+++ user_modules/shared_testcases/testCasesForPolicyTable.lua
@@ -25,10 +25,7 @@ local defaultFunctionGroupName = "group1"
 testCasesForPolicyTable.time_trigger = 0
 testCasesForPolicyTable.time_onstatusupdate = 0
 testCasesForPolicyTable.time_policyupdate = 0
---testCasesForPolicyTable.PTS_elements = {{}}
-testCasesForPolicyTable.preloaded_elements = {}
-testCasesForPolicyTable.pts_elements = {}
-testCasesForPolicyTable.seconds_between_retries = {}
+
 
 ---------------------------------------------------------------------------------------------
 ------------------------------------------ Functions ----------------------------------------
@@ -846,301 +843,9 @@ function testCasesForPolicyTable:flow_PTU_SUCCEESS_EXTERNAL_PROPRIETARY()
     end 
 end
 
-local data_dictionary = 
-{
-	{ name = "module_meta.pt_exchanged_at_odometer_x", elem_required = "required"},
-	{ name = "module_meta.pt_exchanged_x_days_after_epoch", elem_required = "required"},
-	{ name = "module_meta.ignition_cycles_since_last_exchange", elem_required = "required"}	,
-
-	{ name = "module_config.preloaded_pt", elem_required = "optional"},
-	{ name = "module_config.preloaded_date", elem_required = "optional"},
-	{ name = "module_config.exchange_after_x_ignition_cycles", elem_required = "required"},
-	{ name = "module_config.exchange_after_x_kilometers", elem_required = "required"},
-	{ name = "module_config.exchange_after_x_days", elem_required = "required"},
-	{ name = "module_config.timeout_after_x_seconds", elem_required = "required"},
-	{ name = "module_config.endpoints.0x04.default.priority", elem_required = "required"},
-	{ name = "module_config.endpoints.0x04.default.groups", elem_required = "required"},
-	{ name = "module_config.endpoints.0x04.default.AppHMIType", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x04.default.memory_kb", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x04.default.heart_beat_timeout_ms", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x04.default.RequestType", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x07.default.priority", elem_required = "required"},
-	{ name = "module_config.endpoints.0x07.default.groups", elem_required = "required"},
-	{ name = "module_config.endpoints.0x07.default.AppHMIType", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x07.default.memory_kb", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x07.default.heart_beat_timeout_ms", elem_required = "optional"},
-	{ name = "module_config.endpoints.0x07.default.RequestType", elem_required = "optional"},
-	{ name = "module_config.notifications_per_minute_by_priority.EMERGENCY", elem_required = "required"},
-	{ name = "module_config.notifications_per_minute_by_priority.NAVIGATION", elem_required = "required"},
-	{ name = "module_config.notifications_per_minute_by_priority.VOICECOM", elem_required = "required"},
-	{ name = "module_config.notifications_per_minute_by_priority.COMMUNICATION", elem_required = "required"},
-	{ name = "module_config.notifications_per_minute_by_priority.NORMAL", elem_required = "required"},
-	{ name = "module_config.notifications_per_minute_by_priority.NONE", elem_required = "required"},
-	{ name = "module_config.certificate", elem_required = "optional"},
-	{ name = "module_config.vehicle_make", elem_required = "optional"},
-	{ name = "module_config.vehicle_model", elem_required = "optional"},
-	{ name = "module_config.vehicle_year", elem_required = "optional"},
-	{ name = "module_config.display_order", elem_required = "optional"},
-
-	{ name = "consumer_friendly_messages.version", elem_required = "required"},
-
-	{ name = "app_policies.default.priority", elem_required = "required"},
-	{ name = "app_policies.default.groups", elem_required = "required"},
-	{ name = "app_policies.default.AppHMIType", elem_required = "optional"},
-	{ name = "app_policies.default.memory_kb", elem_required = "optional"},
-	{ name = "app_policies.default.heart_beat_timeout_ms", elem_required = "optional"},
-	{ name = "app_policies.default.RequestType", elem_required = "optional"},
-	{ name = "app_policies.pre_DataConsent.priority", elem_required = "required"},
-	{ name = "app_policies.pre_DataConsent.groups", elem_required = "required"},
-	{ name = "app_policies.pre_DataConsent.AppHMIType", elem_required = "optional"},
-	{ name = "app_policies.pre_DataConsent.memory_kb", elem_required = "optional"},
-	{ name = "app_policies.pre_DataConsent.heart_beat_timeout_ms", elem_required = "optional"},
-	{ name = "app_policies.pre_DataConsent.RequestType", elem_required = "optional"}
-
-}
-local json_elements = {}
-
-local function extract_json(pathToFile)
-	json_elements = {}
-	if( commonSteps:file_exists(pathToFile) ) then
-		print("file exist")
-	
-		local file  = io.open(pathToFile, "r")
-		local json_data = file:read("*all") 
-		file:close()
-
-		local json = require("modules/json")
-		local data = json.decode(json_data)
-		local i = 1
-
-		for index_level1, value_level1 in pairs(data.policy_table) do
-			if(type(value_level1) == false) then
-		      json_elements[i] = { name = index_level1 , elem_required = nil, value = value_level1}
-		      i = i + 1
-			else
-				for index_level2, value_level2 in pairs(value_level1) do
-					if( type(value_level2) ~= "table" ) then 
-				      json_elements[i] = { name = index_level1.."."..index_level2, elem_required = nil, value = value_level2 }
-				      i = i + 1
-					else
-						for index_level3, value_level3 in pairs(value_level2) do
-
-							if(type(value_level3) ~= "table") then 
-						      json_elements[i] = { name = index_level1 .. "."..index_level2.."."..index_level3 , elem_required = nil, value = value_level3 }
-						      i = i + 1
-							else
-								for index_level4, value_level4 in pairs(value_level3) do
-									
-									if(type(value_level4) ~= "table") then 
-						      			json_elements[i] = { name = index_level1 .. "."..index_level2 .. "."..index_level3.."."..index_level4, elem_required = nil, value = value_level4 }
-						      			i = i + 1
-						      		else
-						      			for index_level5, value_level5 in pairs(value_level4) do
-						      				if(type(value_level5) ~= "table") then
-						      					json_elements[i] = { name = index_level1 .. "."..index_level2 .. "."..index_level3.. "."..index_level4.."."..index_level5, elem_required = nil, value = value_level5 }
-						      					i = i + 1
-						      				else
-						      					for index_level6, value_level6 in pairs(value_level5) do
-						      						if(type(value_level6) ~= "table") then
-						      							json_elements[i] = { name = index_level1 .. "."..index_level2 .. "."..index_level3.. "."..index_level4  .. "."..index_level5.."."..index_level6, elem_required = nil, value = value_level6 }
-						      							i = i + 1
-						      						end
-						      					end
-						      				end
-						      			end
-									end
-								end
-							end
-						end
-					end
-				end
-		   end
-		end
-	else
-		print("file doesn't exist: " ..pathToFile)
-	end
-end
-
-function testCasesForPolicyTable:extract_preloaded_pt()
-	preloaded_pt = 'SDL_bin/sdl_preloaded_pt.json'	
-	extract_json(preloaded_pt)
-	k = 1
-	for i = 1, #json_elements do
-		testCasesForPolicyTable.preloaded_elements[i] = { name = json_elements[i].name, value = json_elements[i].value }
-		if( string.sub(json_elements[i].name,1,string.len("module_config.seconds_between_retries.")) == "module_config.seconds_between_retries." ) then
-			testCasesForPolicyTable.seconds_between_retries[k] = { name = json_elements[i].name, value = json_elements[i].value}
-			k = k + 1
-		end
-	end
-end
-
-function testCasesForPolicyTable:create_PTS(is_created, app_IDs, device_IDs)
-	if(is_created == false) then
-		if ( commonSteps:file_exists( '/tmp/fs/mp/images/ivsu_cache/sdl_snapshot.json') ) then
-			print(" \27[31m /tmp/fs/mp/images/ivsu_cache/sdl_snapshot.json is created \27[0m")	
-		end
-	else
-		testCasesForPolicyTable:extract_preloaded_pt()
-		local j = 1
-		local k = 1
-		local length_data_dict = #data_dictionary
-		for i = 1, #json_elements do
-			local str_1 = json_elements[i].name
-			if( string.sub(str_1,1,string.len("functional_groupings.")) == "functional_groupings." ) then
-				data_dictionary[length_data_dict + j] = { name = json_elements[i].name, value = json_elements[i].value, elem_required = "required" }
-				j = j + 1
-			end
-
-			if( string.sub(str_1,1,string.len("module_config.seconds_between_retries.")) == "module_config.seconds_between_retries." ) then
-				testCasesForPolicyTable.seconds_between_retries[k] = { name = json_elements[i].name, value = json_elements[i].value}
-				k = k + 1
-				data_dictionary[length_data_dict + j] = { name = json_elements[i].name, value = json_elements[i].value, elem_required = "required" }
-				j = j + 1
-			end
-
-			if( string.sub(str_1,1,string.len("app_policies.pre_DataConsent.groups.")) == "app_policies.pre_DataConsent.groups." ) then
-				data_dictionary[length_data_dict + j] = { name = json_elements[i].name, value = json_elements[i].value, elem_required = "required" }
-				j = j + 1
-			end
-
-			if( string.sub(str_1,1,string.len("app_policies.default.groups.")) == "app_policies.default.groups." ) then
-				data_dictionary[length_data_dict + j] = { name = json_elements[i].name, value = json_elements[i].value, elem_required = "required" }
-				j = j + 1
-			end
-
-			if( string.sub(str_1,1,string.len("module_config.endpoints.0x07.default.")) == "module_config.endpoints.0x07.default." ) then
-				data_dictionary[length_data_dict + j] = { name = json_elements[i].name, value = json_elements[i].value, elem_required = "required" }
-				j = j + 1
-			end
-
-			if( string.sub(str_1,1,string.len("module_config.endpoints.0x04.default.")) == "module_config.endpoints.0x04.default." ) then
-				data_dictionary[length_data_dict + j] = { name = json_elements[i].name, value = json_elements[i].value, elem_required = "required" }
-				j = j + 1
-			end
-		end	
-
-		if(app_IDs ~= nil) then
-			for i = 1, #app_IDs do
-				local length_data_dict = #data_dictionary
-				data_dictionary[length_data_dict + 1] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".nicknames", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 2] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".priority", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 3] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".groups", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 4] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".AppHMIType", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 5] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".memory_kb", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 6] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".heart_beat_timeout_ms", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 7] = { name = "module_config.endpoints."..tostring(app_IDs[i])..".RequestType", value = nil, elem_required = "optional" }
-
-				data_dictionary[length_data_dict + 8] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".count_of_tls_errors", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 9] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".nicknames", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 10] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".priority", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 11] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".groups", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 12] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".AppHMIType", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 13] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".memory_kb", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 14] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".heart_beat_timeout_ms", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 15] = { name = "usage_and_error_counts."..tostring(app_IDs[i])..".RequestType", value = nil, elem_required = "optional" }
-				
-				data_dictionary[length_data_dict + 16] = { name = "app_policies."..tostring(app_IDs[i])..".count_of_tls_errors", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 17] = { name = "app_policies."..tostring(app_IDs[i])..".nicknames", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 18] = { name = "app_policies."..tostring(app_IDs[i])..".priority", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 19] = { name = "app_policies."..tostring(app_IDs[i])..".groups", value = nil, elem_required = "required" }
-				data_dictionary[length_data_dict + 20] = { name = "app_policies."..tostring(app_IDs[i])..".AppHMIType", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 21] = { name = "app_policies."..tostring(app_IDs[i])..".memory_kb", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 22] = { name = "app_policies."..tostring(app_IDs[i])..".heart_beat_timeout_ms", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 23] = { name = "app_policies."..tostring(app_IDs[i])..".RequestType", value = nil, elem_required = "optional" }
-				data_dictionary[length_data_dict + 24] = { name = "app_policies."..tostring(app_IDs[i]), value = nil, elem_required = "optional" }
-			end
-		else
-			data_dictionary[#data_dictionary + 1] = { name = "usage_and_error_counts", elem_required = "required"}
-		end
 
 
-		if(device_IDs ~= nil) then
-			for i =1 , #device_IDs do
-				local length_data_dict = #data_dictionary
-				data_dictionary[length_data_dict + 1] = { name = "device_data."..tostring(device_IDs[i])..".usb_transport_enabled", value = nil, elem_required = "required" }
-			end
-		else
-			data_dictionary[#data_dictionary + 1] = { name = "device_data", elem_required = "required"}
-		end
 
-		pts = '/tmp/fs/mp/images/ivsu_cache/sdl_snapshot.json'
-		if ( commonSteps:file_exists(pts) ) then
-			extract_json(pts)
-			local k = 1
-			for i = 1, #json_elements do
-				testCasesForPolicyTable.pts_elements[k] = {
-					name = json_elements[i].name, 
-					value = json_elements[i].value
-				}
-				k = k + 1
-			end	
-		else
-			print("/tmp/fs/mp/images/ivsu_cache/sdl_snapshot.json doesn't exits! ")
-		end
-		
-		--Check for ommited parameters
-		for i = 1, #testCasesForPolicyTable.pts_elements do
-			local str_1 = testCasesForPolicyTable.pts_elements[i].name
-			local is_existing = false
-			for j = 1, #data_dictionary do
-				local str_2 = data_dictionary[j].name
-				if( str_1 == str_2 ) then
-					is_existing = true
-					for k = 1, #testCasesForPolicyTable.preloaded_elements do
-						if(testCasesForPolicyTable.preloaded_elements[k].name == str_2) then
-							if(testCasesForPolicyTable.pts_elements[i].value ~= testCasesForPolicyTable.preloaded_elements[k].value) then
-								print(testCasesForPolicyTable.pts_elements[i].name .." = " .. tostring(testCasesForPolicyTable.pts_elements[i].value) .. ". Should be " ..tostring(testCasesForPolicyTable.preloaded_elements[k].value) )
-							end
-						end
-					end
-					break
-				end
-			end
-			if (is_existing == false) then
-				print(testCasesForPolicyTable.pts_elements[i].name .. ": should not exist")
-			end
-		end
-
-		--Check for mandatory elements
-		for i = 1, #data_dictionary do
-			if(data_dictionary[i].elem_required == "required") then
-				local str_2 = data_dictionary[i].name
-				local is_existing = false
-				for j = 1, #testCasesForPolicyTable.pts_elements do
-					local str_1 = testCasesForPolicyTable.pts_elements[j].name
-					if( str_1 == str_2 ) then
-						is_existing = true
-						break
-					end
-				end
-				if (is_existing == false) then
-					print(data_dictionary[i].name .. ": mandatory parameter does not exist in PTS")
-				end
-			end
-		end
-	end --if(is_created == false) then
-end
-
-function testCasesForPolicyTable:get_data_from_PTS(pts_element)
-	local value
-	local is_found = false
-	for i = 1, #testCasesForPolicyTable.pts_elements do
-		if (pts_element == testCasesForPolicyTable.pts_elements[i].name) then
-			value = testCasesForPolicyTable.pts_elements[i].value
-			is_found = true
-			break
-		end
-	end		
-	if(is_found == false) then 
-		print(" \27[31m Element "..pts_element.." is not found in PTS! \27[0m")
-	end
-	if (value == nil)	then
-		print(" \27[31m Value of "..pts_element.." is nil \27[0m")
-		value = 0
-	end
-
-	return value
-end
 
 function testCasesForPolicyTable:trigger_PTU_user_request_update_from_HMI()
 	function Test:TestStep_trigger_PTU_user_request_update_from_HMI()
